Внедрен механизм CAPTCHA. CAPTCHA при регистрации юзеров. Фамилию юзеру указывать не обязательно (конфигурируется). Запрет вставлять файлы неизвестного типа в качестве текста или картинки в постинге. Поля creator_id, modifier_id в entries; created, modified в users. Багфиксы в скриптах. Переходим на mt_rand(), srand() не требуется с PHP 4.2.0.

+++ lib/mail.php
@@ -128,9 +128,10 @@ while($row=mysql_fetch_assoc($result))
        $runResult=$sendResult;
      $sentList[]=$row['id'];
      }
-sql('delete from mail_queue
-     where id in ('.join(',',$sentList).')',
-    __FUNCTION__);
+if(count($sentList)>0)
+  sql('delete from mail_queue
+       where id in ('.join(',',$sentList).')',
+      __FUNCTION__);
 unlockMailTables();
 return $runResult;
 }
